refactor: move ManualRunButton
